Move UIViewController, UIWindow categories from StripePaymentSheet to StripeUICore
