+++ StripeUICore/StripeUICore/Source/Categories/UIViewController+StripeUICore.swift
@@ -38,4 +38,16 @@ import UIKit
         }
         return self
     }
+    
+    /// Walks the presented view controller hierarchy and return the top most presented controller.
+    /// - Returns: Returns the top most presented view controller, or `nil` if this view controller is not presenting another controller.
+    func findTopMostPresentedViewController() -> UIViewController? {
+        var topMostController = self.presentedViewController
+
+        while let presented = topMostController?.presentedViewController {
+            topMostController = presented
+        }
+
+        return topMostController
+    }
 }
